+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
-# 0.0.1
+# 0.2.0
+
+## Features:
+
+add posibility to check staled Pull-Request, togheter with Issues.
+
+Right now the bot comment automatically PRs and Issue. 
+
+- add logging and try/catch
+
+- improve layout
+
+# 0.1.0
 
 Add initial structure. Implement old issue check